Show a toast when an unknown media key is pressed
This is just an idea.  If people want more media keys supported, they
will be able to use this toast to report exactly what key codes their
hardware is generating.

If you don't want to include this, just let me know and I can remove it
from the pull request.