Updated QueueFragment to have a refresh button.
Issue #595

Also resolved an issue that may have been latent when 'All Episodes'
was default. Namely, that the search and other option menus didn't
show up until you navigated away and then back to the fragment.
This was happening because the items weren't loaded yet, which prevented
the options menu from being created.  If we switch back to 'All Episodes'
as being default, this problem may appear again.