Fix for AntennaPod/AntennaPod#629
On older devices, crucial intermediate certificates are missing. In
this fix, we create a custom trustmanager. The algorithm follows the
certificate chain and checks validity and issuer signature. In the last
step, we take the root CA's certificate from the system's keystore and
verify the chain's last certificate thus validating the whole chain.