Fixed some NullPointerExceptions
These were happening on some handsets when onPrepareOptionsMenu was getting called.

It is sometimes called before the menu is actually loaded, causing us to not find the
MenuItem we're looking for.

This solves the symptom, if not the cause.

After making this change the number of failures reported on Apkudo dropped from 20 devices to just 6.