Updated version code for RC5
Also updated the oldVersion used by the UPDATE FeedItems sql statement.
We don't want to run the 1030002 version ever, and the 1030005 version
should fix any users that used the 1030002 version (with the possibility
that some items that shouldn't have had auto_download set to 0 did have
it set to 0, but accepting that that's not a terrible problem).