Add dialog to choose lock screen playback buttons
This adds a dialog to choose the playback buttons on the lock screen
notification. It only allows selecting a maximum of two values, because
the lock screen notification can only display up to 3 buttons and the
play/pause button is always included. It defaults to additionally show
the skip button.

The minimum sdk has been changed back to 10.